Move some code to utils in beacon client
