+++ relayer/src/ethereum_beacon_client/mod.rs
@@ -6,17 +6,12 @@ use std::{
 };
 
 use anyhow::{anyhow, bail, Error as AnyError, Result as AnyResult};
-use ark_serialize::CanonicalDeserialize;
-use checkpoint_light_client_io::{
-    ArkScale, BeaconBlockHeader, G1TypeInfo, G2TypeInfo, Slot, SyncCommitteeKeys,
-    SyncCommitteeUpdate, G1, G2, SYNC_COMMITTEE_SIZE,
-};
+use checkpoint_light_client_io::{BeaconBlockHeader, Slot};
 use ethereum_common::{
-    base_types::{BytesFixed, FixedArray},
-    beacon::{BLSPubKey, Block as BeaconBlock},
+    beacon::Block as BeaconBlock,
     utils::{
         BeaconBlockHeaderResponse, BeaconBlockResponse, FinalityUpdate, FinalityUpdateResponse,
-        Update, UpdateResponse,
+        UpdateResponse,
     },
     MAX_REQUEST_LIGHT_CLIENT_UPDATES,
 };
@@ -24,6 +19,7 @@ use reqwest::{Client, ClientBuilder, RequestBuilder};
 use serde::{de::DeserializeOwned, Deserialize};
 
 pub mod slots_batch;
+pub mod utils;
 
 #[derive(Clone, Debug)]
 pub struct ErrorNotFound;
@@ -213,87 +209,6 @@ impl BeaconClient {
             .await
             .map(|response| response.data)
     }
-}
-
-pub fn map_public_keys(
-    compressed_public_keys: &FixedArray<BLSPubKey, SYNC_COMMITTEE_SIZE>,
-) -> Box<SyncCommitteeKeys> {
-    let keys = compressed_public_keys
-        .0
-        .iter()
-        .map(|BytesFixed(pub_key_compressed)| {
-            let pub_key = <G1 as CanonicalDeserialize>::deserialize_compressed_unchecked(
-                &pub_key_compressed.0[..],
-            )
-            .expect("Public keys have the required size");
-
-            let ark_scale: ArkScale<G1TypeInfo> = G1TypeInfo(pub_key).into();
-
-            ark_scale
-        })
-        .collect::<Vec<_>>();
-
-    Box::new(FixedArray(keys.try_into().expect(
-        "The size of keys array is guaranteed on the type level",
-    )))
-}
-
-pub fn sync_update_from_finality(
-    signature: G2,
-    finality_update: FinalityUpdate,
-) -> SyncCommitteeUpdate {
-    SyncCommitteeUpdate {
-        signature_slot: finality_update.signature_slot,
-        attested_header: finality_update.attested_header,
-        finalized_header: finality_update.finalized_header,
-        sync_aggregate: finality_update.sync_aggregate,
-        sync_committee_next_aggregate_pubkey: None,
-        sync_committee_signature: G2TypeInfo(signature).into(),
-        sync_committee_next_pub_keys: None,
-        sync_committee_next_branch: None,
-        finality_branch: finality_update
-            .finality_branch
-            .into_iter()
-            .map(|BytesFixed(array)| array.0)
-            .collect::<_>(),
-    }
-}
-
-pub fn sync_update_from_update(signature: G2, update: Update) -> SyncCommitteeUpdate {
-    let next_sync_committee_keys = map_public_keys(&update.next_sync_committee.pubkeys);
-
-    SyncCommitteeUpdate {
-        signature_slot: update.signature_slot,
-        attested_header: update.attested_header,
-        finalized_header: update.finalized_header,
-        sync_aggregate: update.sync_aggregate,
-        sync_committee_next_aggregate_pubkey: Some(update.next_sync_committee.aggregate_pubkey),
-        sync_committee_signature: G2TypeInfo(signature).into(),
-        sync_committee_next_pub_keys: Some(next_sync_committee_keys),
-        sync_committee_next_branch: Some(
-            update
-                .next_sync_committee_branch
-                .into_iter()
-                .map(|BytesFixed(array)| array.0)
-                .collect::<_>(),
-        ),
-        finality_branch: update
-            .finality_branch
-            .into_iter()
-            .map(|BytesFixed(array)| array.0)
-            .collect::<_>(),
-    }
-}
-
-pub fn try_from_hex_encoded<T: TryFrom<Vec<u8>>>(hex_encoded: &str) -> Option<T> {
-    let data = match hex_encoded.starts_with("0x") {
-        true => &hex_encoded[2..],
-        false => hex_encoded,
-    };
-
-    hex::decode(data)
-        .ok()
-        .and_then(|bytes| <T as TryFrom<Vec<u8>>>::try_from(bytes).ok())
 }
 
 async fn get<R: DeserializeOwned>(request_builder: RequestBuilder) -> AnyResult<R> {
